requirements.txt
fastapi
uvicorn[standard]
pymongo
celery[redis]
google-cloud-storage
python-multipart
python-dotenv
ffmpeg-python

